updated signup user and volunteer, fixed db, added a home page
