+++ app/src/main/java/com/example/thirdeye/user_registration/SignUP.java
@@ -1,0 +1,37 @@
+package com.example.thirdeye.user_registration;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+
+import com.example.thirdeye.R;
+
+public class SignUP extends AppCompatActivity {
+
+    private Button sign_in, sign_up;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_sign_u_p);
+        sign_in = findViewById(R.id.sign_in);
+        sign_up = findViewById(R.id.sign_up);
+        sign_up.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(SignUP.this, SignUpOptions.class);
+                startActivity(intent);
+            }
+        });
+        sign_in.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(SignUP.this, SignInUser.class);
+                startActivity(intent);
+            }
+        });
+    }
+}
