+++ app/src/main/res/layout/activity_home.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:gravity="center"
+    tools:context=".Home">
+
+
+    <Button
+        android:id="@+id/tri_sos"
+        android:layout_width="100dp"
+        android:layout_marginRight="10dp"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:text="SOS" />
+
+    <Button
+        android:id="@+id/tri_video"
+        android:layout_width="100dp"
+        android:layout_marginLeft="10dp"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:text="VIDEO" />
+</LinearLayout>
